+++ packages/react-universal/types/index.d.ts
@@ -1,8 +1,10 @@
 export { ActiveProps } from "./Active";
 export { BlockProps } from "./Block";
 export { RouterProps, CuriRenderFn } from "./curiProvider";
+export { NavigatingProps } from "./Navigating";
 import Active from "./Active";
 import Block from "./Block";
 import curiProvider from "./curiProvider";
 import { Curious } from "./Context";
-export { Active, Block, curiProvider, Curious };
+import Navigating from "./Navigating";
+export { Active, Block, curiProvider, Curious, Navigating };