+++ packages/router/src/curi.ts
@@ -5,7 +5,7 @@ import matchLocation from "./matchLocation";
 import resolveMatchedRoute from "./resolveMatchedRoute";
 import { privatePrepareRoutes } from "./prepareRoutes";
 
-import { History, PendingNavigation } from "@hickory/root";
+import { History, PendingNavigation, Action } from "@hickory/root";
 
 import {
   RouteDescriptor,
@@ -25,7 +25,9 @@ import {
   RemoveObserver,
   CurrentResponse,
   Navigation,
-  NavigationDetails
+  NavigationDetails,
+  Cancellable,
+  RemoveCancellable
 } from "./types/curi";
 
 export default function createRouter(
@@ -62,6 +64,7 @@ export default function createRouter(
 
   let observers: Array<Observer> = [];
   const oneTimers: Array<Observer> = [];
+  let cancellers: Array<Cancellable> = [];
 
   function setupRoutesAndInteractions(userRoutes?: UserRoutes): void {
     if (userRoutes) {
@@ -86,10 +89,7 @@ export default function createRouter(
   }
 
   function navigationHandler(pendingNav: PendingNavigation): void {
-    if (activeNavigation) {
-      activeNavigation.cancel(pendingNav.action);
-      activeNavigation.cancelled = true;
-    }
+    cancelActiveNavigation(pendingNav.action);
     activeNavigation = pendingNav;
 
     const navigation: Navigation = {
@@ -126,6 +126,7 @@ export default function createRouter(
     if (match.route.sync) {
       finalizeResponseAndEmit(match as Match, pendingNav, navigation, null);
     } else {
+      activateCancellers(pendingNav.action);
       resolveMatchedRoute(match as Match, external).then(
         (resolved: ResolveResults) => {
           if (pendingNav.cancelled) {
@@ -148,6 +149,7 @@ export default function createRouter(
     navigation: Navigation,
     resolved: ResolveResults | null
   ) {
+    deactivateCancellers();
     const response = finishResponse(
       match,
       routeInteractions,
@@ -197,6 +199,40 @@ export default function createRouter(
     }
   }
 
+  let canCancel: boolean;
+  function activateCancellers(action: Action) {
+    if (cancellers.length) {
+      canCancel = true;
+      const cancel = () => {
+        cancelActiveNavigation(action);
+        deactivateCancellers();
+        if (cancelCallback) {
+          cancelCallback();
+        }
+        resetCallbacks();
+      };
+      cancellers.forEach(fn => {
+        fn(cancel);
+      });
+    }
+  }
+
+  function deactivateCancellers() {
+    if (canCancel) {
+      canCancel = false;
+      cancellers.forEach(fn => {
+        fn();
+      });
+    }
+  }
+
+  function cancelActiveNavigation(action: Action) {
+    if (activeNavigation) {
+      activeNavigation.cancel(action);
+      activeNavigation.cancelled = true;
+    }
+  }
+
   const router: CuriRouter = {
     route: routeInteractions,
     history,
@@ -221,6 +257,14 @@ export default function createRouter(
       } else {
         oneTimers.push(fn);
       }
+    },
+    cancel(fn: Cancellable): RemoveCancellable {
+      cancellers.push(fn);
+      return () => {
+        cancellers = cancellers.filter(can => {
+          return can !== fn;
+        });
+      };
     },
     refresh(routes?: Array<RouteDescriptor>) {
       refreshing = true;