+++ website/src/pages/Packages/router.js
@@ -598,6 +598,40 @@ stopObserving();
                 </Section>
               </Section>
 
+              <Section tag="h5" title="cancel(fn)" id="cancel-property">
+                <Explanation>
+                  <p>
+                    With asynchronous routes, after a user begins navigation,
+                    but before the route's asynchronous actions have finished,
+                    the user does not have a good way to cancel the navigation.
+                    They can either refresh the page (causing a full reload) or
+                    click a link with the same URL as the current location, but
+                    neither of these are intuitive or ideal.
+                  </p>
+                  <p>
+                    <IJS>cancel()</IJS> takes an observer function that will be
+                    called when navigation starts and when the navigation is
+                    finished. When the navigation starts, the observer function
+                    will be given a function to cancel the navigation. When the
+                    navigation finishes, the function will be called with{" "}
+                    <IJS>undefined</IJS>.
+                  </p>
+                  <p>
+                    Calling <IJS>cancel()</IJS> returns a function to stop
+                    observing.
+                  </p>
+                </Explanation>
+                <CodeBlock>
+                  {`const stopCancelling = router.cancel(fn => {
+  if (fn === undefined) {
+    // the navigation has finished/been cancelled
+  } else {
+    // calling fn will cancel the navigation
+  }
+});`}
+                </CodeBlock>
+              </Section>
+
               <Section tag="h5" title="current()" id="current-property">
                 <Explanation>
                   <p>